project/kafka-services/consumer-trend/requirements.txt
pandas
psycopg2-binary
confluent-kafka
